cloud_function/requirements.txt
kfp==1.6.2
google-cloud-aiplatform
google-cloud-storage
